travis+jbuilder debugging

+++ .travis-ci.sh
@@ -29,7 +29,7 @@ init-bootstrap () {
     eval $(opam env)
     # extlib is installed, since UChar.cmi causes problems with the search
     # order. See also the removal of uChar and uTF8 in src_ext/jbuild-extlib-src
-    opam install cohttp-lwt-unix ssl cmdliner dose3 opam-file-format re extlib 'jbuilder>=1.0+beta19' 'mccs>=1.1+5' --yes
+    opam install cohttp-lwt-unix ssl cmdliner dose3 cudf.0.9 opam-file-format re extlib 'jbuilder>=1.0+beta19' 'mccs>=1.1+5' --yes
   fi
   rm -f "$OPAMBSROOT"/log/*
 }
@@ -147,6 +147,7 @@ export OCAMLRUNPARAM=b
     make lib-ext
   fi
   make all
+  if [ $? -ne 0 ]; then cat boot.ml; exit 1; fi
 
   rm -f ~/local/bin/opam
   make install
